feat: 라이브러리 기능 전반 요청 및 응답 구조 리팩토링 및 DTO 분리 적용 - 전체/상세 조회 DTO 분리 - 등록 API 응답 DTO 수정 - summaryText, video metadata 포함 - 식별자 데이터 타입 Long으로 모두 변경 - dummy.init 설정 기반 더미데이터 삽입 방식 개선

+++ src/main/java/com/kdt/yts/YouSumback/model/entity/Question.java
@@ -6,7 +6,7 @@ import jakarta.persistence.*;
 public class Question {
     @Id
     @Column(name = "question_id", nullable = false)
-    private int questionId; // 질문 식별자
+    private Long questionId; // 질문 식별자
 
     @ManyToOne
     @JoinColumn(name = "quiz_id", nullable = false)
